refactor: extract formatting helpers to shared utils for reuse
