+++ src/lib/utils/date.ts
@@ -1,0 +1,7 @@
+export function formatWorkoutDate(timestamp: number) {
+  return new Intl.DateTimeFormat(undefined, {
+    month: 'short',
+    day: 'numeric',
+    year: 'numeric'
+  }).format(new Date(timestamp));
+}
